Unsupervised Learning - All Life Bank Customer Segmentation
Unsupervised Learning - All Life Bank Customer Segmentation
Context:-

All Life Bank wants to focus on its credit card customer base in the next financial year. They have been advised by their marketing research team, that the penetration in the market can be improved. Based on this input, the marketing team proposes to run personalized campaigns to target new customers as well as upsell to existing customers.

Another insight from the market research was that the customers perceive the support services of the bank poorly. Based on this, the operations team wants to upgrade the service delivery model, to ensure that customers' queries are resolved faster. The head of marketing and the head of delivery, both decide to reach out to the Data Science team for help.

Objective:-

Identify different segments in the existing customer base, taking into account their spending patterns as well as past interactions with the bank.